+++ cmd/kobs/plugins/plugins.go
@@ -14,6 +14,7 @@ import (
 	"github.com/kobsio/kobs/plugins/applications"
 	"github.com/kobsio/kobs/plugins/dashboards"
 	"github.com/kobsio/kobs/plugins/elasticsearch"
+	"github.com/kobsio/kobs/plugins/flux"
 	"github.com/kobsio/kobs/plugins/jaeger"
 	"github.com/kobsio/kobs/plugins/kiali"
 	"github.com/kobsio/kobs/plugins/markdown"
@@ -33,9 +34,10 @@ type Config struct {
 	Prometheus    prometheus.Config    `json:"prometheus"`
 	Elasticsearch elasticsearch.Config `json:"elasticsearch"`
 	Jaeger        jaeger.Config        `json:"jaeger"`
-	Markdown      markdown.Config      `json:"markdown"`
 	Kiali         kiali.Config         `json:"kiali"`
+	Flux          flux.Config          `json:"flux"`
 	Opsgenie      opsgenie.Config      `json:"opsgenie"`
+	Markdown      markdown.Config      `json:"markdown"`
 	RSS           rss.Config           `json:"rss"`
 }
 
@@ -68,9 +70,10 @@ func Register(clusters *clusters.Clusters, config Config) chi.Router {
 	router.Mount(prometheus.Route, prometheus.Register(clusters, router.plugins, config.Prometheus))
 	router.Mount(elasticsearch.Route, elasticsearch.Register(clusters, router.plugins, config.Elasticsearch))
 	router.Mount(jaeger.Route, jaeger.Register(clusters, router.plugins, config.Jaeger))
-	router.Mount(markdown.Route, markdown.Register(clusters, router.plugins, config.Markdown))
 	router.Mount(kiali.Route, kiali.Register(clusters, router.plugins, config.Kiali))
+	router.Mount(flux.Route, flux.Register(clusters, router.plugins, config.Flux))
 	router.Mount(opsgenie.Route, opsgenie.Register(clusters, router.plugins, config.Opsgenie))
+	router.Mount(markdown.Route, markdown.Register(clusters, router.plugins, config.Markdown))
 	router.Mount(rss.Route, rss.Register(clusters, router.plugins, config.RSS))
 
 	return router